refactor(web): replace hand-rolled fetch client with Eden Treaty
Phase 5b of the Elysia migration: replace the custom fetch client
(~285 lines of auth refresh, timeout, rate limit machinery) with
Eden Treaty backed by a custom fetcher.

New:
- web/src/api/eden.ts — Eden treaty client with custom fetcher
  (auth refresh on 401, 10s timeout, rate limit extraction)
- web/src/api/routes.ts — typed API functions moved from
  shared/api.ts, rewritten to use Eden proxy chain

Deleted:
- web/src/api/client.ts — replaced by Eden custom fetcher

Trimmed:
- shared/api.ts — runtime functions removed, type-only now
  (481 lines → 120 lines of interface/type exports)

Net: 771 lines deleted, 114 added (reduction of 657 lines)
Tests: 306 pass, 0 fail.
