+++ oxlint.config.ts
@@ -636,7 +636,7 @@ export default defineConfig({
     // Web client files: console is the debug/error logger for browser-side code
     {
       files: [
-        'packages/web/src/api/client.ts',
+        'packages/web/src/api/eden.ts',
         'packages/web/src/context/AuthContext.tsx',
         'packages/web/src/components/settings/CompressorSection.tsx',
         'packages/web/src/components/settings/EqualizerSection.tsx',
@@ -644,6 +644,24 @@ export default defineConfig({
       ],
       rules: {
         'no-console': 'off',
+      },
+    },
+
+    // Eden-backed API routes: tsgo cannot resolve Eden's deeply-nested proxy
+    // types. Bun's transpiler handles them correctly. no-unsafe-* warnings
+    // are inherent until tsgo supports Elysia's generic types.
+    // .then(unwrap) is the cleanest pattern for unwrapping { data, error }.
+    {
+      files: ['packages/web/src/api/routes.ts'],
+      rules: {
+        '@typescript-eslint/no-explicit-any': 'off',
+        '@typescript-eslint/no-unsafe-assignment': 'off',
+        '@typescript-eslint/no-unsafe-member-access': 'off',
+        '@typescript-eslint/no-unsafe-call': 'off',
+        '@typescript-eslint/no-unsafe-return': 'off',
+        '@typescript-eslint/no-unsafe-argument': 'off',
+        'promise/prefer-await-to-then': 'off',
+        'unicorn/no-useless-undefined': 'off',
       },
     },
 
@@ -684,6 +702,8 @@ export default defineConfig({
         'packages/web/src/components/VirtualSongList.tsx',
         'packages/web/src/pages/PlaylistDetailPage.tsx',
         'packages/web/src/pages/SongsPage.tsx',
+        // Eden fetcher: retried guard variable for 401 refresh loop prevention
+        'packages/web/src/api/eden.ts',
       ],
       rules: {
         '@typescript-eslint/no-unnecessary-condition': 'off',
@@ -698,7 +718,6 @@ export default defineConfig({
         'packages/server/src/shared/api.ts',
         'packages/server/src/shared/shuffle.ts',
         'packages/server/src/lib/jwt.ts',
-        'packages/web/src/api/client.ts',
         'packages/web/src/hooks/useSocket.ts',
       ],
       rules: {
@@ -803,7 +822,7 @@ export default defineConfig({
         'packages/server/src/GuildPlayer.ts',
         'packages/server/src/lib/syncPlaylistToTag.ts',
         'packages/server/src/lib/voice.ts',
-        'packages/web/src/api/client.ts',
+        'packages/web/src/api/eden.ts',
         'packages/web/src/pages/PlaylistDetailPage.tsx',
         'packages/server/src/lib/tagCanonicalization.ts',
         'packages/server/src/lib/migrateExistingTags.ts',
