+++ packages/web/package.json
@@ -14,6 +14,7 @@
     "@atlaskit/pragmatic-drag-and-drop": "^2.0.1",
     "@atlaskit/pragmatic-drag-and-drop-auto-scroll": "^3.0.0",
     "@atlaskit/pragmatic-drag-and-drop-hitbox": "^2.0.0",
+    "@elysia/eden": "^1.4.10",
     "@phosphor-icons/react": "2.1.10",
     "@tanstack/react-virtual": "3.14.8",
     "masonic": "^4.1.0",
