+++ packages/web/src/components/BulkEditModal.tsx
@@ -1,7 +1,8 @@
-import { type BulkEditData, fetchTags, type TagItem } from '@alfira/server/shared/api';
+import { type BulkEditData, type TagItem } from '@alfira/server/shared/api';
 import { EraserIcon, XIcon } from '@phosphor-icons/react';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
+import { fetchTags } from '../api/api';
 import { useTagColors } from '../context/TagsContext';
 import { getTagColorClasses } from '../utils/tagColors';
 import { Backdrop } from './Backdrop';
