+++ packages/web/src/utils/api.ts
@@ -1,4 +1,4 @@
-import { ApiError } from '../api/client';
+import { ApiError } from '../api/eden';
 
 /** Extract error message from a Fetch-based API error, with a fallback. */
 export function apiErrorMessage(err: unknown, fallback: string): string {
